requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
cachelib==0.10.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.2
comm==0.1.2
cryptography==40.0.2
debugpy==1.6.6
decorator==5.1.1
distlib==0.3.4
executing==1.2.0
filelock==3.6.0
Flask==2.2.2
Flask-OAuthlib==0.9.6
Flask-SQLAlchemy==3.0.3
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-oauth==1.0.1
greenlet==2.0.2
idna==3.4
ipykernel==6.21.1
ipython==8.9.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.0.2
jupyter_core==5.2.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
npm==0.1.1
numpy==1.24.2
oauth==1.0.1
oauthlib==2.1.0
optional-django==0.1.0
packaging==23.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.14.0
PyMySQL==1.0.3
pyOpenSSL==23.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pyzmq==25.0.0
requests==2.28.2
requests-oauthlib==1.1.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.9
stack-data==0.6.2
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.14
virtualenv==20.14.1
wcwidth==0.2.6
Werkzeug==2.2.2
